refactor: extract diagnostics publisher
